+++ docs/ChangeLog.md
@@ -6,6 +6,7 @@
 
 - **Feature**: Ability to export a return as an excel file
 - **Update**: Don't add sample fixtures when adding a new MCA
+- **Update**: Order admin MCA user list (admin first, otherwise ordered by name)
 
 ## 7th July 2025